main_server/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0
yfinance==0.2.36
python-dotenv==1.0.0
websockets==12.0
pymysql==1.1.0
